Add support for rounding to end of sentence
Includes initial tests (all passing), though could do with more?

If options.round is useds, then as the text is parsed, if it hits the character limit, it also checks to see if it's at the edge of a closing sentence tag . If not, it'll continue to scoop up the text.

If the parser reaches a PARSER_TAG_COMMENCED character (`<`), then it peeks at the upcoming HTML, and checks if that tag would suggest the end of a sentence, i.e. a </p>, </div>, </li>, etc.